Disable -w flag for CUDA on Visual Studio
Because the CUDA builds were experiencing some warnings about unused
functions, a -w flag was added to our nvcc flags. This effectively
turned off the warning (which was part of -Wall). But on Visual Studio
this added flag was causing a warning of its own because it overwrote
the /W3 flag. So turn this off for Visual Studio.

I don't think Visual Studio gives this warning anyway. If it does, we
can add a specific flat to suppress that warning.