Allow clang-format to pass more empty lines
Previously, if you had more than 1 empty line somewhere, clang-format
would eat that empty line. But it is a bit of a pain sometimes to get
the autogenerated code to have the exact empty lines, and there are
valid reasons to allow more than one empty line somewhere, so relax this
condition a bit.