Fix compiler errors for UnitTestMapperGLFW
This file was creating a C array using the size of an array handle.
Generally, you cannot create a C array of a static size with a
variable created at runtime. Apparently some compilers were able
to trace back to fact that the array handle was a given size and
used that, but not all compilers can do this. To get around the
problem, just use a std::vector, which allows the data to be
dynamically allocated.