Fix Marching Cubes regression test
Turns out there was a bug in the way Marching Cubes computed gradients
on a structured grid. A recent change to Boundary also had a similar bug
that ended in the same behavior. The recent changes to BoundaryState
fixed the problem, but caused the Marching Cubes test to fail because it
was comparing against bad gradient values.

This change fixes the array of gradient values. I'm pretty sure these
are correct since I generated them with ParaView (using a completely
separate implementation).